DEFAULT: modulate affine registration tolerance
I have noticed that the initial affine registration does not properly align the template to some subject images, so that the following diffeo+affine starts in a sub-optimal alignment. Modulating the affine tolerance seems to solve this. It does increase runtime, but not by much, as if is only applied to the coarsest level of the algorithm.